Cleaning up, pass 3 (breaking)
This removes the split API provided for generic `IsString` embedding and `ByteString` specific embedding, because `ByteString`s are already instances of `IsString`, and we  (both library developers and library consumers) can just let GHC take care of the gruntwork involved.

The injection support is not something that will be used by all (or most) users of the library, but it is close enough to the purpose of the library that it is still included, but in a submodule now.